# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Documentazione GDR Node
  version: 1.0.0
  description: API Documentation of our gdr
paths:
...
  /class:
    get:
      summary: Ottieni tutte le classi
      tags:
        - Classes
      responses:
        "200":
          description: Lista di tutte le classi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
        "500":
          description: Errore interno del server
    post:
      summary: Crea una nuova classe
      tags:
        - Classes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Classe creata con successo
        "400":
          description: Dati non validi
        "500":
          description: Errore interno del server
  "/class/{id}":
    get:
      summary: Ottieni una classe per ID
      tags:
        - Classes
      parameters:
        - in: path
          name: id
          required: true
          description: ID della classe
          schema:
            type: string
      responses:
        "200":
          description: Dettagli della classe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Classe non trovata
        "500":
          description: Errore interno del server
    patch:
      summary: Aggiorna una classe esistente
      tags:
        - Classes
      parameters:
        - in: path
          name: id
          required: true
          description: ID della classe da aggiornare
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Class"
      responses:
        "200":
          description: Classe aggiornata con successo
        "400":
          description: Dati non validi
        "404":
          description: Classe non trovata
        "500":
          description: Errore interno del server
    delete:
      summary: Elimina una classe
      tags:
        - Classes
      parameters:
        - in: path
          name: id
          required: true
          description: ID della classe da eliminare
          schema:
            type: string
      responses:
        "200":
          description: Classe eliminata con successo
        "404":
          description: Classe non trovata
        "500":
          description: Errore interno del server
...
components:
  schemas:
...
    Class:
      type: object
      required:
        - name
        - key
        - description
      properties:
        id:
          type: string
          description: L'ID univoco della classe
        name:
          type: string
          description: Il nome della classe
        key:
          type: string
          description: La chiave univoca della classe
        description:
          type: string
          description: Descrizione della classe
        createdAt:
          type: string
          format: date-time
          description: La data di creazione della classe
        updatedAt:
          type: string
          format: date-time
          description: La data dell'ultimo aggiornamento della classe
    ClassInput:
      type: object
      required:
        - name
        - key
        - description
      properties:
        name:
          type: string
          description: Il nome della classe
        key:
          type: string
          description: La chiave univoca della classe
        description:
          type: string
          description: Descrizione della classe
        createdAt:
          type: string
          format: date-time
          description: La data di creazione della classe
        updatedAt:
          type: string
          format: date-time
          description: La data dell'ultimo aggiornamento della classe
...